Aha! Bash line wrapping/PS1 working again.
This is an old case that seems to be cracked (=understood by me). All
non-printable characters of a bash prompt must be enclosed in '\[' and
'\]', if not they will be counted when calculating the length of the
prompt, hence screw up the line wrapping when you are typing a long
enough command. Classic case of copy-paste without understanding the
details.